Allow BookScreen playback in background
